refactor: organize results of LocalStorageService
