+++ lib/src/core/services/local_storage/results/local_storage_results.dart
@@ -1,0 +1,27 @@
+final class SuccessSaveData {}
+
+final class FailureSaveData {}
+
+final class SuccessReadData {
+  Object data;
+
+  SuccessReadData({required this.data});
+}
+
+final class FailureReadData {
+  String message;
+
+  FailureReadData({required this.message});
+}
+
+final class SuccessDeleteAll {
+  bool isDeleted = false;
+
+  SuccessDeleteAll({required this.isDeleted});
+}
+
+final class FailureDeleteAll {
+  String message;
+
+  FailureDeleteAll({required this.message});
+}
